refactored AddButton (split inputs into separate files in List/AddButton/
