+++ client/src/Components/ReviewsContainer/Breakdowns/ProductBkdn.jsx
@@ -5,11 +5,6 @@ import Col from "react-bootstrap/Col"
 import Form from "react-bootstrap/Form"
 import useReviews from "../../../contexts/ReviewsContext.jsx"
 
-const scaleStyle = {
-  display: "flex",
-  justifyContent: "space-between",
-}
-
 const ProductBkdn = () => {
   const { reviewMetadata, details } = useReviews()
 
@@ -33,12 +28,7 @@ const ProductBkdn = () => {
     <Container className="productBkdn">
       {reviewMetadata.characteristics && details
         ? Object.keys(reviewMetadata.characteristics).map((char, i) => (
-            <Scale
-              key={i}
-              char={char}
-              {...reviewMetadata.characteristics[char]}
-              style={scaleStyle}
-            />
+            <Scale key={i} char={char} {...reviewMetadata.characteristics[char]} />
           ))
         : null}
     </Container>
